+++ src/components/Navbar.tsx
@@ -4,10 +4,8 @@ import {
 } from "mdb-react-ui-kit";
 import { Link } from "react-router-dom";
 import { useContext } from "react";
-import { LivesContext } from "../hooks/UseLives";
 
 const NavbarPage: React.FC = () => {
-  const {lives} = useContext(LivesContext);
   return (
       <MDBNavbar color="dark" dark expand="md">
         <MDBNavbarBrand>
@@ -15,8 +13,6 @@ const NavbarPage: React.FC = () => {
             <strong className="white-text" style={{marginLeft: '20px'}}>MeleeGuessr</strong>
           </Link>
         </MDBNavbarBrand>
-        
-        <span className='ms-auto navbar-text dark-red-text' style={{marginRight: '20px'}}>Lives: {lives}</span>
       </MDBNavbar>
     );
   }