+++ src/consts/Colors.ts
@@ -1,4 +1,4 @@
-export const Colors = ['primary', 'secondary', 'success', 'danger', 'warning', "info"]
+export const Colors = ['primary', 'secondary', 'warning', "info"]
 
 export const randomColor = () => {
   return Colors[Math.round(Math.random() * Colors.length)]